+++ src/bk-login/bklogin/authentication/urls.py
@@ -9,11 +9,16 @@ an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express o
 specific language governing permissions and limitations under the License.
 """
 from django.urls import path
+from django.views.generic import TemplateView
 
-from . import views
+from . import api_views, views
 
 urlpatterns = [
+    # 登录入口
     path("", views.LoginView.as_view()),
+    # 前端页面（选择登录的用户）
+    path("pages/users", TemplateView.as_view(template_name="index.html")),
+    # ------------------------------------------ 租户 & 登录方式选择 ------------------------------------------
     # 租户配置
     path("tenant-global-settings/", views.TenantGlobalSettingRetrieveApi.as_view()),
     # 租户信息
@@ -23,18 +28,18 @@ urlpatterns = [
     path("sign-in-tenants/", views.SignInTenantCreateApi.as_view()),
     # 认证源
     path("idps/", views.TenantIdpListApi.as_view()),
-    # 已认证后的用户
+    # ------------------------------------------ 认证插件 ------------------------------------------
+    # 插件认证
+    path("auth/idps/<str:idp_id>/actions/<str:action>/", views.IdpPluginDispatchView.as_view()),
+    # ------------------------------------------ 用户选择 ------------------------------------------
+    # 已认证后的用户列表
     path("tenant-users/", views.TenantUserListApi.as_view()),
     # 确认登录的用户
     path("sign-in-users/", views.SignInTenantUserCreateApi.as_view()),
-    # Test
-    path("index/", views.IndexView.as_view()),
-    # API
-    path("api/v1/is_login/", views.CheckTokenApi.as_view()),
-    path("api/v1/get_user/", views.GetUserApi.as_view()),
 ]
 
+# OpenAPI
 urlpatterns += [
-    # 各种认证
-    path("auth/idps/<str:idp_id>/actions/<str:action>/", views.IdpPluginDispatchView.as_view()),
+    path("api/v1/is_login/", api_views.CheckTokenApi.as_view()),
+    path("api/v1/get_user/", api_views.GetUserApi.as_view()),
 ]
